Restore original package structure for platform files
Move HttpClientProvider back to core/network, FileExporter to core/export,
and AppDetailsProvider to app/data to restore original package organization
while keeping codecov exclusion via **/platform/** pattern.
