+++ shared/src/androidMain/kotlin/space/be1ski/vibits/shared/core/network/HttpClientProvider.kt
@@ -1,7 +1,7 @@
-package space.be1ski.vibits.shared.core.platform
+package space.be1ski.vibits.shared.core.network
 
 import io.ktor.client.HttpClient
-import io.ktor.client.engine.cio.CIO
+import io.ktor.client.engine.okhttp.OkHttp
 import io.ktor.client.plugins.contentnegotiation.ContentNegotiation
 import io.ktor.client.plugins.logging.LogLevel
 import io.ktor.client.plugins.logging.Logger
@@ -11,10 +11,10 @@ import io.ktor.serialization.kotlinx.json.json
 import kotlinx.serialization.json.Json
 
 /**
- * Desktop-specific HTTP client configuration.
+ * Android-specific HTTP client configuration.
  */
 actual fun createHttpClient(): HttpClient =
-  HttpClient(CIO) {
+  HttpClient(OkHttp) {
     install(ContentNegotiation) {
       json(
         Json {
