+++ shared/src/commonMain/kotlin/space/be1ski/vibits/shared/app/data/Exporter.kt
@@ -2,8 +2,8 @@ package space.be1ski.vibits.shared.app.data
 
 import dev.zacsweers.metro.Inject
 import kotlinx.serialization.json.Json
+import space.be1ski.vibits.shared.core.export.FileExporter
 import space.be1ski.vibits.shared.core.logging.Log
-import space.be1ski.vibits.shared.core.platform.FileExporter
 import space.be1ski.vibits.shared.feature.memos.data.offline.OfflineMemosFileDto
 import space.be1ski.vibits.shared.feature.memos.data.platform.OfflineMemoStorage
 import kotlin.time.Clock
