+++ shared/src/commonMain/kotlin/space/be1ski/vibits/shared/app/domain/usecase/LoadAppDetailsUseCase.kt
@@ -1,8 +1,8 @@
 package space.be1ski.vibits.shared.app.domain.usecase
 
 import dev.zacsweers.metro.Inject
+import space.be1ski.vibits.shared.app.data.AppDetailsProvider
 import space.be1ski.vibits.shared.app.domain.model.AppDetails
-import space.be1ski.vibits.shared.core.platform.AppDetailsProvider
 
 /**
  * Loads app details for settings screen.
