+++ shared/src/iosMain/kotlin/space/be1ski/vibits/shared/core/export/FileExporter.kt
@@ -1,4 +1,4 @@
-package space.be1ski.vibits.shared.core.platform
+package space.be1ski.vibits.shared.core.export
 
 import kotlinx.cinterop.BetaInteropApi
 import kotlinx.cinterop.ExperimentalForeignApi
