+++ shared/src/wasmJsMain/kotlin/space/be1ski/vibits/shared/core/network/HttpClientProvider.kt
@@ -1,20 +1,18 @@
-package space.be1ski.vibits.shared.core.platform
+package space.be1ski.vibits.shared.core.network
 
 import io.ktor.client.HttpClient
-import io.ktor.client.engine.okhttp.OkHttp
+import io.ktor.client.engine.js.Js
 import io.ktor.client.plugins.contentnegotiation.ContentNegotiation
 import io.ktor.client.plugins.logging.LogLevel
-import io.ktor.client.plugins.logging.Logger
 import io.ktor.client.plugins.logging.Logging
-import io.ktor.http.HttpHeaders
 import io.ktor.serialization.kotlinx.json.json
 import kotlinx.serialization.json.Json
 
 /**
- * Android-specific HTTP client configuration.
+ * Creates a JS [HttpClient] configured for the Memos API.
  */
 actual fun createHttpClient(): HttpClient =
-  HttpClient(OkHttp) {
+  HttpClient(Js) {
     install(ContentNegotiation) {
       json(
         Json {
@@ -24,13 +22,6 @@ actual fun createHttpClient(): HttpClient =
       )
     }
     install(Logging) {
-      logger =
-        object : Logger {
-          override fun log(message: String) {
-            println("HTTP | $message")
-          }
-        }
       level = LogLevel.INFO
-      sanitizeHeader { it == HttpHeaders.Authorization }
     }
   }
